Rename `demo_RTL.dart` to `demo_rtl.dart`
